improve: migrate filters to TypeScript and enhance implementations
