--- src/playback/filters/dsp/clamp16Bit.js
@@ -1,3 +1,0 @@
-export function clamp16Bit(sample) {
-  return Math.max(-32768, Math.min(32767, Math.round(sample)))
-}
