+++ src/playback/processing/filtersManager.ts
@@ -12,23 +12,23 @@ import type {
 import type { FiltersState } from '../../typings/playback/player.types.ts'
 import type { IFiltersManager } from '../../typings/playback/processing.types.ts'
 
-import ChannelMix from '../filters/channelMix.js'
-import Chorus from '../filters/chorus.js'
-import Compressor from '../filters/compressor.js'
-import Distortion from '../filters/distortion.js'
-import Echo from '../filters/echo.js'
-import Equalizer from '../filters/equalizer.js'
-import Flanger from '../filters/flanger.js'
-import Highpass from '../filters/highpass.js'
-import Karaoke from '../filters/karaoke.js'
-import Lowpass from '../filters/lowpass.js'
-import Phaser from '../filters/phaser.js'
-import Reverb from '../filters/reverb.js'
-import Rotation from '../filters/rotation.js'
-import Spatial from '../filters/spatial.js'
-import Timescale from '../filters/timescale.js'
-import Tremolo from '../filters/tremolo.js'
-import Vibrato from '../filters/vibrato.js'
+import ChannelMix from '../filters/channelMix.ts'
+import Chorus from '../filters/chorus.ts'
+import Compressor from '../filters/compressor.ts'
+import Distortion from '../filters/distortion.ts'
+import Echo from '../filters/echo.ts'
+import Equalizer from '../filters/equalizer.ts'
+import Flanger from '../filters/flanger.ts'
+import Highpass from '../filters/highpass.ts'
+import Karaoke from '../filters/karaoke.ts'
+import Lowpass from '../filters/lowpass.ts'
+import Phaser from '../filters/phaser.ts'
+import Reverb from '../filters/reverb.ts'
+import Rotation from '../filters/rotation.ts'
+import Spatial from '../filters/spatial.ts'
+import Timescale from '../filters/timescale.ts'
+import Tremolo from '../filters/tremolo.ts'
+import Vibrato from '../filters/vibrato.ts'
 
 const FILTER_CLASSES: Record<string, FilterClass> = {
   tremolo: Tremolo,
